renamed commitChart folder
